Major fix to GroupByGLA. Added ExtremeTuplesGLA. Formatting changes to some files.
GroupByGLA previously did not call Finalize() on all of its member GLAs when
being finalized itself. This has been corrected.

ExtremeTuplesGLA added. This GLA collects all tuples that have a ranking set of
attributes equal to the extreme of those attributes in that set of data
(i.e. maximum or minimum, specified per attribute). If there is more than
one attribute in the ranking set, one tuple being "better" than another is
compared in order of first ranking attribute to last (i.e. the first
ranking attribute is most significant).

Some formatting changes were made to some files. This mostly includes removal
of excess whitespace (whitespace at the end of a line) and changing tabs to
spaces. In particular, significant formatting changes were made to the
Join LHS work function and WPLib in order to make the generated code easier
to read and look more like it was generated by hand.

Some minor changes were made to the svn:ignore property of some directories
to remove files that are generated during compilation or at runtime.

+++ trunk/src/Column/headers/HStringIterator.h
@@ -68,14 +68,14 @@ public:
 	// is taken out of the iterator and put into iterateMe
 	void Done (Column &iterateMe);
 
-	// advance to the next object in the column... 
+	// advance to the next object in the column...
 	void Advance ();
 
 
 	// add a new data object into the column at the current position, overwriting the
 	// bytes that are already there.  Note that if the size of addMe differs from the
 	// size of the object that is already there, addMe will over-run part of the next
-	// object.  Sooooo... overwrite existing objects in the column with care! 
+	// object.  Sooooo... overwrite existing objects in the column with care!
 	void Insert (HString &addMe);
 	void Insert (const HString &addMe);
 
@@ -83,7 +83,7 @@ public:
 	// it should not be read (it is undefined what happens if you read it)
 	int AtUnwrittenByte ();
 
-	// returns the data object at the current position in the column... 
+	// returns the data object at the current position in the column...
 	HString GetCurrent ();
 
 	// Is the string frequent
@@ -127,7 +127,7 @@ HStringIterator& HStringIterator::operator=(const HStringIterator& other) {
 	// any crap can be here because this is just to support STL
 }
 
-inline 
+inline
 HString HStringIterator :: GetCurrent () {
 
 	if (it.IsInvalid ())
@@ -138,13 +138,13 @@ HString HStringIterator :: GetCurrent () {
 	return HString(*(__uint64_t*)myData, *((__uint64_t*)myData + 1), myData+16);
 }
 
-inline 
+inline
 int HStringIterator :: AtUnwrittenByte () {
 	//return (objLen == 0);
 	return it.AtUnwrittenByte ();
 }
 
-inline 
+inline
 bool HStringIterator :: IsFrequent (__uint64_t hashVal) {
 
 	totalCNT++;
@@ -155,9 +155,9 @@ bool HStringIterator :: IsFrequent (__uint64_t hashVal) {
 
 		return (it->second > MIN_FREQUENT) || (it->second > MIN_THRESHOLD*totalCNT);
 	}
-	
+
 	hashFrequencyMap[hashVal]=1;
-	
+
 	return false;
 }
 
@@ -207,7 +207,7 @@ void HStringIterator :: Insert (HString &addMe) {
     *((__uint64_t*) myData + 1) = addMe.GetStrLength();
     // GetString() is sure to have some value if string is not in dictionary
     strcpy(myData + 16, addMe.GetString());
-	FATALIF(addMe.GetStrLength() > 256, "String is too weird size %ld", addMe.GetStrLength());
+	FATALIF(addMe.GetStrLength() > 256, "String is too weird size %lld", (long long) addMe.GetStrLength());
   }
 }
 
@@ -226,7 +226,7 @@ void HStringIterator :: Advance () {
     it.SetObjLen (((HString*) it.GetData())->GetObjLength ());
 
   //fprintf(stderr, "\nSize string = %ld", it.GetObjLen());
-	FATALIF(it.GetObjLen() > 256, "String is too weird size %ld", it.GetObjLen());
+	FATALIF(it.GetObjLen() > 256, "String is too weird size %lld", (long long) it.GetObjLen());
   it.EnsureSpace (it.GetObjLen(), it.GetObjLen());
 }
